Merge functions in the environment selection page
In the environment selection page, one option is to create an
environment by setting it's name and the path to the configuration file
and another option is to create it by setting it's name and selecting
the file in a list.

Now that multiple input can use the same name and that the <form>'s ID
is no more prefixed to the <input>'s names, the function to check and
apply both forms are identical.