Allow adding path to PYTHONPATH from task file
When an external project declares a task and some tools (python function
that can be called by a requirement), the running Flask process and it's
current PYTHONPATH will not reach the task dedicated tools.

This patch allows the task to list the directories, relatively to the
task file or using an absolute path, with the tools.

The running instance of Flask, when it will load the task file, will
read these directories and add them to the current PYTHONPATH.