Add note to check behavior with a list in the requirements
In one hand, the function db_tools.get_values returns a list and in
another hand, if an element is an array, the function
requirements.solve_element consider that the function need to
recursively solve the content of the array.

It would be nice to have an automated test that uses the
db_tools.get_values function and makes sure that the result is not
considered as an element to solve.