Clean tests for class Task
Add a private function in the test_task class to get a task using the
default task collection.

A part of the test_task tests is about creating a task and these tests
were not modified.

The rest of the tests, on another hand, is about what happens after the
task is created so there is no need to itemize each steps of creating
that task.