Fix volatile collection database 'add_value' method
The volatile collection database was inheriting from the volatile
key/value database only to implement the overriden functions load and
save but were not inheriting functions related to the collection
database.

When calling from a CollectionVolatileDatabase object a function related
to the CollectionDatabase class, it raises this kind of error:

AttributeError: 'CollectionVolatileDatabase' object has no attribute
'add_value'

Inheriting the KV db is a bad idea because the Collection db will have
the functions dedicated to KV db and not those for the collection db.

This commit introduces a CommonVolatileDatabase class with the load and
save function and both KV db and Collection DB will inherit this
class. Then each database (key/value and collection) inherits this class
specific to a volatile storage and the class specific to it's database.