Support 'any' and 'each' in dependencies
Until now, the dependencies were a list of requirement's identifier and
all of these requirement's should be true for the dependencies to be
considered as fulfilled.

Now, the dependencies can be a single requirement's identifier or a
list.

In case of a list, the first element should be "each" or "any" and the
other elements should be requirement's identifier.

If the first element of the list is "any", at least one of the
references requirement's must be true for the dependencies to be
considered fulfilled. If the firt element is "each" then all of them
must be true.

An element can, in turn, be a list itself.