Add support of "safe automated checks"
Prior to this patch, the dependencies needed to be met before checking
if a requirement is fulfiled or not.

This is critical because sometimes, checking a requirements needs
prerequisites. For example, a requirement could check if a property is
set, the property name being set in a dependency, running a check leads
to checking an unknown property.

Sometimes though, the dependencies are irrelevant. This is the case when
the test checks that the property named after the requirement's name is
set. If a dependency is not met but the property is indeed set, there is
no point in considering the requirement's as not met.

Moreover, this tools is about letting the user take back the control and
forcing values in case of necessity but it doesn't work if the
dependencies are blocking. For example, a requirement is about setting a
value and an automatic resolution requires a different property to be
set. If the requirement's property is manually set, there is no need to
check the dependency's property, we can discard the requirement and move
on.

Now, the requirement can declare a "safe_auto_check" and describe
an automated check that doesn't rely on the dependencies. If this test
passes, the requirement will be considered as fulfilled without checking
the dependnecies status nor the "auto_check" test.