Do perform auto resolution even if req is fulfilled
It was decided that if the requirement's self check indicates that the
requirement is already fulfilled, the automatic resolution should not be
applied.

This may be a stupid idea since most of the resolutions are quick and if
the automatic resolution depends on other requirement's and that the
requirement changes we may want to replay the resolution.