Better display of manual resolution
The description of a manual resolution can use this kind of formatting:

```
This is text using a [database|env.db_key] value
```

When displaying this description on the flask GUI, if "db_key" is set in
the environment key/value database, it will be replaced by it's
value. If the key is not set then the text before the pipe will be
disaplyed instead.