Fix IDs issue with macro 'render_input_text()'
The macro 'render_input_text()' used a <form> ID and the name to
construct a unique identifier used for the <intput>'s ID and name but
both properties have different usage and in some cases, the name may be
used in multiple forms.

To fix this issue, the name property doesn't use the <form>'s ID
anymore.

Moreover, the macro now receives an extra parameter related to the
<form>'s submit button and set the <input>'s default value only if it's
related to the <form> that was submitted.