Add a test for class Task
The test checks that the automatic check is not performed if
thedependencies are not met.

Indeed the automatic steps (checks or resolutions) may require data set
by the dependencies.

For example, an automatic check can get the current branch of a
repository and for this it would require the path to the repository to
be set in a key/value database. If the automatic check was performed
with this dependency not met it would run a git command with no path to
the repository and it would print unwanted errors.