GUI: enhance collapsible blocks
The id of the collapsible element should be unique and wether it should
be displayed by default or not doesn't necessarily depends on this id.

Now, the collapsible macro uses a new argument to determine if the
content should be displayed when the page is loaded.

It's content will be displayed id:

* This new argument is set to True;
* This new argument is a parameter of the GET request;

The first case is to force the content to be displayed, unrelated to the
current request.

The second case is to open a collapsible element of a submitted form.