Make function MMMMove function to check dependencies
The function to check if a requirement's dependencies are met is now
part of the Task class API.

This funciton was part of the method to check a requirement's status. A
requirement can be ready if it's automatic check returns true and if all
of it's dependencies are met.

In the case where we want to apply an automatic resolution, the
automatic check will return false but the dependencies should be met
because the dependent requirements can impact the resolution.

Making this function available to anyone will allow a "user context"
(probably an UI) to check if the requirement's is ready to apply the
automatic resolution or not.