Add "safe" property to auto resolutions
This allows to execute automatic resolutions even if some dependencies
are not met.

By default a resolution is unsafe.

Resolutions should not be considered safe without meticulous study to
prevent running automatic resolutions that relies on other requirements
to be filled. This could lead to a simple Python exception not catched
to the unsexpected file removal.