Allow collapsible elements in the tree view
The main problem in the tree view is that there are a lot of elements
and it is difficult to understand where we are on the resolution
process.

This patch add the possibility to specify some requirements as
collapsible.

In this case, when the requirement is displayed:

* It's dependencies are not displayed;
* It's link redirects to it's sub-tree and not it's resolutions.

This causes multiple not-desired behaviors:

* The 'Get back' link redirects to the root tree;
* The 'tree_id' reference is no more reliable;

This should be addressed in a near futur.