Mark the dead-end requirements in the GUI
There are steps that only checks if a requirement is reached. These
steps doesn't really have resolutions. Their only purpose is to run the
automatic resolution procedure.

They can, however, have resolutions configured but this is only to
document the requirement for the user to understand it's purpose.

In this case, the Flask GUI will display a tag "[x]" after their label
to notify the user that it is useless to open it's page to find
automatic or assisted resolutions.