Fix GUI tree conflict with collapsed and dead-end
When a requirement is tagged 'collapsed', the GUI displays '>' in the
tree page. On another hand when a requirement is detected as a
'dead-end' (no automatic nor manual resolutions), the GUI displays
'[x]'.

If a requirement falls into both properties, the CSS rules will not
display both marks.

In this case, the fact that the requirement is collapsed is considered
as more important. This patch makes sure the '>' mark is displayed in
this case.