Fix redirection after failed resolution
When a resolution is used, automated or manual, and the modification
succeeds, the user is redirected to the same page he was on instead of
simply rendering the page. This allows to remove the action from the
current URL. Otherwise, the form data would still be there and
refreshing the page would re-run the resolution.

When an error was encountered while running the resolution, the user was
not redirected. This is now fixed.

Another issue that this patch resolves is breadcrumb parameter in the
tree page after a failed quick-access resolution. The 'prev' parameter
contained the attempted resolution so the page of the attempted
resolution was displayed instead of it's parent.