Fix quick select/copy in Flask GUI
If the user wants to select a value proposed in a hint (manual
resolution), he can triple-click on the text and copy/paste it in the
input area.

The problem is that on Firefox (and maybe others), the triple-click was
selecting white characters before and after. Removing the line ends and
spaces between some tags did not fixed it. Adding a <div> tag before and
after without newlines nor whitespaces solves this issue.